+++ modules/text/module_text_llm/module_text_llm/helpers/feedback_icl/store_feedback_icl.py
@@ -4,7 +4,7 @@ import os
 from athena.text import Submission, Exercise, Feedback
 from typing import List
 from athena.logger import logger
-from module_text_llm.generate_embeddings import embed_bert
+from module_text_llm.helpers.feedback_icl.generate_embeddings import embed_bert
 from module_text_llm.helpers.feedback_icl.store_indices_icl import store_embedding_index
 """
 We store the reference of the feedback in the index file.
@@ -13,16 +13,11 @@ We also need to store the index with the actual feedback now.
 def store_feedback_icl(submission: Submission, exercise: Exercise, feedbacks: List[Feedback]):
     logger.info("Storing feedback for submission %d of exercise %d.", submission.id, exercise.id)
     for feedback in feedbacks:
-        logger.info("Getting reference")
         chunk = get_reference(feedback, submission.text)
-        logger.info("Chunk: %s", chunk)
         embedding = embed_bert(chunk) 
         save_embedding(embedding, exercise.id)
         store_embedding_index(exercise.id, submission.id, feedback)
         
-
-
-    # index.add(np.array([submission.embedding], dtype=np.float32))
 def save_embedding(embedding, exercise_id):
     if os.path.exists(f"embeddings_{exercise_id}.index"):
         index = faiss.read_index(f"embeddings_{exercise_id}.index")
@@ -32,7 +27,7 @@ def save_embedding(embedding, exercise_id):
     index.add(np.array([embedding], dtype=np.float32))
     faiss.write_index(index, f"embeddings_{exercise_id}.index")
     
-def query_embedding(query_embedding,exercise_id, k=3):
+def query_embedding(query_embedding,exercise_id, k=5):
     index_file = os.path.abspath(f"embeddings_{exercise_id}.index")
 
     
@@ -68,3 +63,6 @@ def get_reference(feedback, submission_text):
         return submission_text[feedback.index_start:feedback.index_end ]
     else:
         return submission_text
+    
+def check_if_embedding_exists(exercise_id):
+    return os.path.exists(f"embeddings_{exercise_id}.index")
