+++ modules/text/module_text_llm/module_text_llm/helpers/format_rag_context.py
@@ -1,7 +1,7 @@
 from athena.logger import logger
 from module_text_llm.index_storage import retrieve_embedding_index, retrieve_feedbacks
 from module_text_llm.storage_embeddings import query_embedding
-from module_text_llm.generate_embeddings import embed_text,embed_bert
+from module_text_llm.helpers.feedback_icl.generate_embeddings import embed_text,embed_bert
 
 def retrieve_rag_context(submission,exercise_id):
     query_submission= embed_bert(submission.text)
