+++ modules/text/module_text_llm/module_text_llm/icl_rag_approach/generate_suggestions.py
@@ -16,6 +16,7 @@ from module_text_llm.icl_rag_approach.prompt_generate_suggestions import Assessm
 from module_text_llm.helpers.feedback_icl.retrieve_rag_context_icl import retrieve_rag_context_icl
 from module_text_llm.icl_rag_approach.agent import TutorAgent
 from module_text_llm.icl_rag_approach.ollama_prompt import system_message_segment, human_message_segment, Segmentation, system_message, human_message
+from module_text_llm.helpers.feedback_icl.store_feedback_icl import check_if_embedding_exists
 async def generate_suggestions(exercise: Exercise, submission: Submission, config:ApproachConfig, debug: bool, is_graded :bool) -> List[Feedback]:
     model = config.model.get_model()  # type: ignore[attr-defined]
     isOllama = isinstance(model, ChatOllama)
@@ -62,26 +63,28 @@ async def generate_suggestions(exercise: Exercise, submission: Submission, confi
         return []
     
     if(isOllama):
+        embeddings_exist = check_if_embedding_exists(exercise.id)
         
-        segmentation_prompt = get_chat_prompt_with_formatting_instructions(
-        model=model, 
-        system_message=system_message_segment, 
-        human_message=human_message_segment, 
-        pydantic_object=Segmentation
-            )
-        segmentation_prompt_input = {
-            "grading_instructions": format_grading_instructions(exercise.grading_instructions, exercise.grading_criteria),
-            "submission": submission.text,
-            "problem_statement": exercise.problem_statement or "No problem statement.",
-        }
-        chain = segmentation_prompt | model 
-        segments = chain.invoke( segmentation_prompt_input)
-        print(segments)
-        for segment in segments:
-            print(segment)
-            print(type(segment))
-            formatted_rag_context += retrieve_rag_context_icl(segment[0],exercise.id)
-        prompt_input["rag_context"] = formatted_rag_context
+        if embeddings_exist:
+            segmentation_prompt = get_chat_prompt_with_formatting_instructions(
+            model=model, 
+            system_message=system_message_segment, 
+            human_message=human_message_segment, 
+            pydantic_object=Segmentation
+                )
+            segmentation_prompt_input = {
+                "grading_instructions": format_grading_instructions(exercise.grading_instructions, exercise.grading_criteria),
+                "submission": submission.text,
+                "problem_statement": exercise.problem_statement or "No problem statement.",
+            }
+            chain = segmentation_prompt | model 
+            segments = chain.invoke( segmentation_prompt_input)
+            
+            for segment in segments:
+                formatted_rag_context += retrieve_rag_context_icl(segment[0],exercise.id)
+            prompt_input["rag_context"] = formatted_rag_context
+        else:
+            prompt_input["rag_context"] = "There are no submission at the moment"
                 
         result = await predict_and_parse(
             model=model,
